+++ docs/source/dev_documentation/changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Added
+
+- `Scenario` clone command
+
 ## 2.1.0 - 2024-03-01
 
 ## Added